Separated Log formatter to own file.
