+++ sokannonser/customlogging/__init__.py
@@ -1,0 +1,1 @@
+from .narval_log_formatter import *
